Message Passing between source/backend.
This closes #47 as we now have good enough specc.
It may still vary in the future tho.

See src/core/auth/Auth.md for this specc.

+++ src/core/auth/Auth.cpp
@@ -2,17 +2,17 @@
 
 zmqpp::message &operator<<(zmqpp::message &msg, const Leosac::Auth::SourceType &t)
 {
-    static_assert(std::is_same<std::underlying_type<Leosac::Auth::SourceType>::type, uint32_t>::value,
+    static_assert(std::is_same<std::underlying_type<Leosac::Auth::SourceType>::type, uint8_t>::value,
             "Bad underlying type for enum");
-    msg << static_cast<uint32_t>(t);
+    msg << static_cast<uint8_t>(t);
     return msg;
 }
 
 zmqpp::message &operator>>(zmqpp::message &msg, Leosac::Auth::SourceType &t)
 {
-    static_assert(std::is_same<std::underlying_type<Leosac::Auth::SourceType>::type, uint32_t>::value,
+    static_assert(std::is_same<std::underlying_type<Leosac::Auth::SourceType>::type, uint8_t>::value,
             "Bad underlying type for enum");
-    uint32_t tmp;
+    uint8_t tmp;
     msg >> tmp;
     t = static_cast<Leosac::Auth::SourceType>(tmp);
     return msg;
@@ -20,17 +20,17 @@ zmqpp::message &operator>>(zmqpp::message &msg, Leosac::Auth::SourceType &t)
 
 zmqpp::message &operator<<(zmqpp::message &msg, const Leosac::Auth::AccessStatus &st)
 {
-    static_assert(std::is_same<std::underlying_type<Leosac::Auth::AccessStatus>::type, uint32_t>::value,
+    static_assert(std::is_same<std::underlying_type<Leosac::Auth::AccessStatus>::type, uint8_t>::value,
             "Bad underlying type for enum");
-    msg << static_cast<uint32_t>(st);
+    msg << static_cast<uint8_t>(st);
     return msg;
 }
 
 zmqpp::message &operator>>(zmqpp::message &msg, Leosac::Auth::AccessStatus &st)
 {
-    static_assert(std::is_same<std::underlying_type<Leosac::Auth::AccessStatus>::type, uint32_t>::value,
+    static_assert(std::is_same<std::underlying_type<Leosac::Auth::AccessStatus>::type, uint8_t>::value,
             "Bad underlying type for enum");
-    uint32_t tmp;
+    uint8_t tmp;
     msg >> tmp;
     st = static_cast<Leosac::Auth::AccessStatus>(tmp);
     return msg;
